+++ resources/views/pages/Weather.blade.php
@@ -1,5 +1,5 @@
 <x-page-template bodyClass='dh-shell bg-gray-200' :SEO="$SEO">
-    <x-shell.nav active="me" />
+    <x-shell.nav active="weather" />
     
     
     <main class="main-content position-relative max-height-vh-100 h-100 border-radius-lg ">
@@ -34,7 +34,7 @@
 
     
         <!-- Navbar -->
-        <x-shell.header title="Dive Trips" />
+        <x-shell.header title="Marine forecast" />
         <!-- End Navbar -->
         <div class="container-fluid py-4">